IR: Add GetNZCV/SetNZCV opcodes
These correspond directly to msr/mrs.

We can also set NZCV to an immediate value by running ccmp with a
condition that will always fail (in our case, that register 0 does not
equal register 0). Compared to an immediate move and an msr, llvm-mca
claims this is the same cycle count but fewer instructions and uops so
probably still a win.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>