IR: Changes Select operation to explicit sized operations
This instruction behaviour was very opaque because the size of the
destination was implicitly calculated while the size of the compares
where given explicitly.

Changes both to be explicit and all of the code the requires working
around it.

Quite a large number of changes since select is used all over the place.
Resulting asm output changes is fairly minimal.